% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanGenes.R
\name{cleanGenes}
\alias{cleanGenes}
\title{Clean genes}
\usage{
cleanGenes(geneList)
}
\arguments{
\item{geneList}{A list containing the entrez gene IDs for all of the
genes found in each article.}
}
\value{
A tibble. The first column is the gene name and
the second column is the number of times that gene is mentioned.
}
\description{
Converts the entrez gene ids into the approved gene symbol and
counts the number of times each gene is mentioned.
}
\examples{
pmids <- scrapeIDs(dataBase = 'pubmed',
                   term = '(vivax malaria[MeSH Terms]) AND (folic acid antagonists[MeSH Terms])')

geneids <- extractGenes(IDs = pmids,
                        nCores = 2)

geneNames <- cleanGenes(geneList = geneids)

}
