% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeGenes.R
\name{scrapeGenes}
\alias{scrapeGenes}
\title{Scrape Genes}
\usage{
scrapeGenes(IDs, nCores = 2, nTries = 5)
}
\arguments{
\item{IDs}{A list of article IDs from the NCBI website.}

\item{nCores}{An integer for the number of cores to use to mine the NCBI articles.}

\item{nTries}{An integer for the number of times to attempt to connect to the NCBI website if it is unable to connect.}
}
\value{
A tibble. The first column is the gene name and
the second column is the number of times that gene is mentioned.
}
\description{
Mines the genes mentioned in each of the articles from \code{scrapeIDs}.
}
\examples{
pmids <- scrapeIDs (dataBase = 'pubmed',
                    term = '(vivax malaria[MeSH]) AND (folic acid antagonists[MeSH])')

geneNames <- scrapeGenes (IDs = pmids,
                          nCores = 2,
                          nTries = 5)

}
