% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeIDs.R
\name{scrapeIDs}
\alias{scrapeIDs}
\title{Scrape IDs}
\usage{
scrapeIDs(dataBase = "pubmed", term = "prostate cancer[MeSH]")
}
\arguments{
\item{dataBase}{A string. Indicates the database from which you want to search.}

\item{term}{A string. The terms used to search the specified database.}
}
\value{
A list containing all of the ids of the articles that matched the search criteria.
}
\description{
Mines article IDs from NCBI databases according to the seach criteria.
}
\examples{
pmids2017 <- scrapeIDs(dataBase = 'pubmed',
                       term = '(prostate cancer[MeSH Terms]) AND 2017[Date - Publication]')

head(pmids2017)

}
