% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenes.R
\name{extractGenes}
\alias{extractGenes}
\title{Extract Genes}
\usage{
extractGenes(IDs, nCores = 2, nTries = 5)
}
\arguments{
\item{IDs}{A list of article ids from the NCBI website.}

\item{nCores}{An integer for the number of cores to use to mine the NCBI articles.}

\item{nTries}{An integer for the number of times to attempt to connect to the NCBI website.}
}
\value{
A list. Each element in the list is also a list. It contains the
entrez gene ids for all the genes found in each abstract.
}
\description{
Extracts the gene ids from the information scraped from the PubTator website.
}
\examples{
pmids <- scrapeIDs(dataBase = 'pubmed',
                   term = '(vivax malaria[MeSH Terms]) AND (folic acid antagonists[MeSH Terms])')

geneNames <- extractGenes(IDs = pmids,
                          nCores = 2,
                          nTries = 5)

}
